+++ src/fr/gardoll/ace/controller/pump/InterfacePousseSeringue.java
@@ -21,7 +21,7 @@ import fr.gardoll.ace.controller.core.InitializationException ;
 //TODO: singleton.
 public class InterfacePousseSeringue  implements Closeable, PumpController
 {
-  private static final int OPENING_DELAY = 2000;
+  private static final long OPENING_DELAY = 2000l;
   
   private static final Pattern _DELIVER_PATTERN = Pattern.compile("([0-9.]+)\\s+(u|m)l\\s+(:|>|<)");
   private static final Pattern _NA_PATTERN = Pattern.compile("NA\\s+(:|>|<)");
